mainApp/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
Werkzeug==3.1.3
